requirements.txt
openai
langchain
python-dotenv
requests
beautifulsoup4
pandas
